refactor: move setup outside versioned sdk
