+++ pkg/plugin/system/aspectplugins.go
@@ -4,7 +4,7 @@ Copyright © 2021 Aspect Build Systems Inc
 Not licensed for re-use.
 */
 
-package v1alpha1
+package system
 
 import (
 	"fmt"
